Fix for model lightmapping param
I made an oopsie woopsy and forgot to actually make the shader use the lightmap texture. This should fix that issue... I also moved some stuff into the else statement for whether or not the model flag is set. That way the shader won't check it when you are using a model... ( wvt things can't be used on models... )